Reduce include nesting a bit
