+++ basilisk/basilisk_DEX.c
@@ -471,51 +471,6 @@ char *basilisk_respond_accept(struct supernet_info *myinfo,bits256 privkey,uint3
         retstr = clonestr("{\"error\":\"couldnt find to requestid to choose\"}");
     return(retstr);
 }
-
-#include "../includes/iguana_apidefs.h"
-#include "../includes/iguana_apideclares.h"
-
-THREE_STRINGS_AND_DOUBLE(tradebot,aveprice,comment,base,rel,basevolume)
-{
-    double retvals[4],aveprice; cJSON *retjson = cJSON_CreateObject();
-    aveprice = instantdex_avehbla(myinfo,retvals,base,rel,basevolume);
-    jaddstr(retjson,"result","success");
-    jaddnum(retjson,"aveprice",aveprice);
-    jaddnum(retjson,"avebid",retvals[0]);
-    jaddnum(retjson,"bidvol",retvals[1]);
-    jaddnum(retjson,"aveask",retvals[2]);
-    jaddnum(retjson,"askvol",retvals[3]);
-    return(jprint(retjson,1));
-}
-
-ZERO_ARGS(InstantDEX,allcoins)
-{
-    struct iguana_info *tmp; cJSON *native,*notarychains,*basilisk,*virtual,*full,*retjson = cJSON_CreateObject();
-    full = cJSON_CreateArray();
-    native = cJSON_CreateArray();
-    basilisk = cJSON_CreateArray();
-    virtual = cJSON_CreateArray();
-    notarychains = cJSON_CreateArray();
-    HASH_ITER(hh,myinfo->allcoins,coin,tmp)
-    {
-        if ( coin->FULLNODE < 0 )
-            jaddistr(native,coin->symbol);
-        //else if ( coin->virtualchain != 0 )
-        //    jaddistr(virtual,coin->symbol);
-        else if ( coin->FULLNODE > 0 )//|| coin->VALIDATENODE > 0 )
-            jaddistr(full,coin->symbol);
-        //else if ( coin->notarychain >= 0 )
-        //    jaddistr(notarychains,coin->symbol);
-        else jaddistr(basilisk,coin->symbol);
-    }
-    jadd(retjson,"native",native);
-    jadd(retjson,"basilisk",basilisk);
-    jadd(retjson,"full",full);
-    //jadd(retjson,"virtual",virtual);
-    //jadd(retjson,"notarychains",notarychains);
-    return(jprint(retjson,1));
-}
-
 cJSON *basilisk_unspents(struct supernet_info *myinfo,struct iguana_info *coin,char *coinaddr)
 {
     cJSON *unspents=0,*array=0; char *retstr;
@@ -558,201 +513,3 @@ char *basilisk_sendrawtransaction(struct supernet_info *myinfo,struct iguana_inf
     return(retstr);
 }
 
-STRING_ARG(InstantDEX,available,source)
-{
-    uint64_t total = 0; int32_t i,n=0; char coinaddr[64]; cJSON *item,*unspents,*retjson = 0;
-    if ( source != 0 && source[0] != 0 && (coin= iguana_coinfind(source)) != 0 )
-    {
-        if ( myinfo->expiration != 0 )
-        {
-            bitcoin_address(coinaddr,coin->chain->pubtype,myinfo->persistent_pubkey33,33);
-            if ( (unspents= basilisk_unspents(myinfo,coin,coinaddr)) != 0 )
-            {
-                //printf("available.(%s)\n",jprint(unspents,0));
-                if ( (n= cJSON_GetArraySize(unspents)) > 0 )
-                {
-                    for (i=0; i<n; i++)
-                    {
-                        item = jitem(unspents,i);
-                        //if ( is_cJSON_True(jobj(item,"spendable")) != 0 )
-                            total += jdouble(item,"amount") * SATOSHIDEN;
-                        //printf("(%s) -> %.8f\n",jprint(item,0),dstr(total));
-                    }
-                }
-                free_json(unspents);
-            }
-            retjson = cJSON_CreateObject();
-            jaddnum(retjson,"result",dstr(total));
-            printf(" n.%d total %.8f (%s)\n",n,dstr(total),jprint(retjson,0));
-            return(jprint(retjson,1));
-        }
-        printf("InstantDEX_available: need to unlock wallet\n");
-        return(clonestr("{\"error\":\"need to unlock wallet\"}"));
-    }
-    printf("InstantDEX_available: %s is not active\n",source!=0?source:"");
-    return(clonestr("{\"error\":\"specified coin is not active\"}"));
-}
-
-HASH_ARRAY_STRING(InstantDEX,request,hash,vals,hexstr)
-{
-    uint8_t serialized[512]; bits256 privkey; char buf[512],BTCaddr[64],KMDaddr[64]; struct basilisk_request R; int32_t jumblr,iambob,optionhours; cJSON *reqjson; uint32_t datalen=0,DEX_channel; struct iguana_info *bobcoin,*alicecoin;
-    myinfo->DEXactive = (uint32_t)time(NULL) + 3*BASILISK_TIMEOUT + 60;
-    jadd64bits(vals,"minamount",jdouble(vals,"minprice") * jdouble(vals,"amount") * SATOSHIDEN * SATOSHIDEN);
-    if ( jobj(vals,"desthash") == 0 )
-        jaddbits256(vals,"desthash",hash);
-    jadd64bits(vals,"satoshis",jdouble(vals,"amount") * SATOSHIDEN);
-    jadd64bits(vals,"destsatoshis",jdouble(vals,"destamount") * SATOSHIDEN);
-    jaddnum(vals,"timestamp",time(NULL));
-    if ( (jumblr= jint(vals,"usejumblr")) != 0 )
-        privkey = jumblr_privkey(myinfo,BTCaddr,KMDaddr,jumblr == 1 ? JUMBLR_DEPOSITPREFIX : "");
-    else privkey = myinfo->persistent_priv;
-    hash = curve25519(privkey,curve25519_basepoint9());
-    if ( jobj(vals,"srchash") == 0 )
-        jaddbits256(vals,"srchash",hash);
-    printf("service.(%s)\n",jprint(vals,0));
-    memset(&R,0,sizeof(R));
-    if ( basilisk_request_create(&R,vals,hash,juint(vals,"timestamp")) == 0 )
-    {
-        iambob = bitcoin_coinptrs(hash,&bobcoin,&alicecoin,R.src,R.dest,privkey,GENESIS_PUBKEY);
-        if ( (optionhours= jint(vals,"optionhours")) != 0 )
-        {
-            printf("iambob.%d optionhours.%d R.requestid.%u vs calc %u, q.%u\n",iambob,R.optionhours,R.requestid,basilisk_requestid(&R),R.quoteid);
-            if ( iambob != 0 && optionhours > 0 )
-            {
-                sprintf(buf,"{\"error\":\"illegal call option request hours.%d when iambob.%d\"}",optionhours,iambob);
-                printf("ERROR.(%s)\n",buf);
-                return(clonestr(buf));
-            }
-            else if ( iambob == 0 && optionhours < 0 )
-            {
-                sprintf(buf,"{\"error\":\"illegal put option request hours.%d when iambob.%d\"}",optionhours,iambob);
-                printf("ERROR.(%s)\n",buf);
-                return(clonestr(buf));
-            }
-        }
-        //if ( myinfo->IAMNOTARY != 0 || myinfo->NOTARY.RELAYID >= 0 )
-        //    R.relaybits = myinfo->myaddr.myipbits;
-        if ( (reqjson= basilisk_requestjson(&R)) != 0 )
-            free_json(reqjson);
-        datalen = basilisk_rwDEXquote(1,serialized,&R);
-        //int32_t i; for (i=0; i<sizeof(R); i++)
-        //    printf("%02x",((uint8_t *)&R)[i]);
-        printf(" R.requestid.%u vs calc %u, q.%u datalen.%d\n",R.requestid,basilisk_requestid(&R),R.quoteid,datalen);
-        basilisk_rwDEXquote(0,serialized,&R);
-    } else printf("error creating request\n");
-    if ( datalen > 0 )
-    {
-        uint32_t msgid;//,crc=0,crcs[2],numiters = 0; uint8_t buf[4096];
-        memset(hash.bytes,0,sizeof(hash));
-        msgid = (uint32_t)time(NULL);
-        DEX_channel = 'D' + ((uint32_t)'E' << 8) + ((uint32_t)'X' << 16);
-        basilisk_channelsend(myinfo,hash,hash,DEX_channel,msgid,serialized,datalen,60);
-        sleep(3);
-        /*while ( numiters < 10 && (crc= basilisk_crcsend(myinfo,0,buf,sizeof(buf),hash,myinfo->myaddr.persistent,DEX_channel,msgid,serialized,datalen,crcs)) == 0 )
-        {
-            //printf("didnt get back what was sent\n");
-            sleep(3);
-            basilisk_channelsend(myinfo,myinfo->myaddr.persistent,hash,DEX_channel,msgid,serialized,datalen,60);
-            numiters++;
-        }*/
-        //if ( crc != 0 )//basilisk_channelsend(myinfo,R.srchash,R.desthash,DEX_channel,(uint32_t)time(NULL),serialized,datalen,30) == 0 )
-            return(clonestr("{\"result\":\"DEX message sent\"}"));
-        //else return(clonestr("{\"error\":\"DEX message couldnt be sent\"}"));
-    }
-    return(clonestr("{\"error\":\"DEX message not sent\"}"));
-}
-
-INT_ARG(InstantDEX,automatched,requestid)
-{
-    // return quoteid
-    myinfo->DEXactive = (uint32_t)time(NULL) + INSTANTDEX_LOCKTIME;
-    return(clonestr("{\"result\":\"automatched not yet\"}"));
-}
-
-int32_t InstantDEX_incoming_func(struct supernet_info *myinfo,void *ptr,uint8_t *data,int32_t datalen)
-{
-    //int32_t i;
-    //for (i=0; i<datalen; i++)
-    //    printf("%02x",data[i]);
-    //printf(" <- incoming\n");
-    return(0);
-}
-
-int32_t InstantDEX_process_channelget(struct supernet_info *myinfo,void *ptr,int32_t (*internal_func)(struct supernet_info *myinfo,void *ptr,uint8_t *data,int32_t datalen),uint32_t channel,uint32_t msgid,uint8_t *data,int32_t datalen,uint32_t expiration,uint32_t duration)
-{
-    return((*internal_func)(myinfo,ptr,data,datalen));
-}
-
-INT_ARG(InstantDEX,incoming,requestid)
-{
-    cJSON *retjson,*retarray; bits256 zero; uint32_t DEX_channel,msgid,now; int32_t retval,width,drift=3; uint8_t data[32768];
-    now = (uint32_t)time(NULL);
-    memset(&zero,0,sizeof(zero));
-    width = (now - myinfo->DEXpoll) + 2*drift;
-    if ( width < (drift+1) )
-        width = 2*drift+1;
-    else if ( width > 64 )
-        width = 64;
-    myinfo->DEXpoll = now;
-    retjson = cJSON_CreateObject();
-    DEX_channel = 'D' + ((uint32_t)'E' << 8) + ((uint32_t)'X' << 16);
-    msgid = (uint32_t)time(NULL) + drift;
-    if ( (retarray= basilisk_channelget(myinfo,zero,myinfo->myaddr.persistent,DEX_channel,msgid,width)) != 0 ) 
-    {
-        //printf("GOT.(%s)\n",jprint(retarray,0));
-        if ( (retval= basilisk_process_retarray(myinfo,0,InstantDEX_process_channelget,data,sizeof(data),DEX_channel,msgid,retarray,InstantDEX_incoming_func)) > 0 )
-        {
-            jaddstr(retjson,"result","success");
-        } else jaddstr(retjson,"error","cant process InstantDEX retarray");
-        jadd(retjson,"responses",retarray);
-    }
-    else
-    {
-        jaddstr(retjson,"error","cant do InstantDEX channelget");
-        printf("error channelget\n");
-    }
-    return(jprint(retjson,1));
-}
-
-/*TWO_INTS(InstantDEX,swapstatus,requestid,quoteid)
-{
-    cJSON *vals; char *retstr;
-    myinfo->DEXactive = (uint32_t)time(NULL) + INSTANTDEX_LOCKTIME;
-    //if ( myinfo->IAMLP != 0 )
-    //    return(basilisk_respond_swapstatus(myinfo,myinfo->myaddr.persistent,requestid,quoteid));
-    //else
-    {
-        vals = cJSON_CreateObject();
-        jaddnum(vals,"requestid",(uint32_t)requestid);
-        jaddnum(vals,"quoteid",(uint32_t)quoteid);
-        jaddbits256(vals,"hash",myinfo->myaddr.persistent);
-        retstr = basilisk_standardservice("SWP",myinfo,0,myinfo->myaddr.persistent,vals,"",1);
-        free_json(vals);
-        return(retstr);
-    }
-}*/
-
-TWO_INTS(InstantDEX,accept,requestid,quoteid)
-{
-    cJSON *vals; char *retstr;
-    myinfo->DEXactive = (uint32_t)time(NULL) + INSTANTDEX_LOCKTIME;
-    if ( myinfo->IAMLP != 0 || myinfo->dexsock >= 0 || myinfo->subsock >= 0 )
-        return(basilisk_respond_accept(myinfo,myinfo->persistent_priv,requestid,quoteid,&myinfo->DEXaccept));
-    else
-    {
-        vals = cJSON_CreateObject();
-        jaddnum(vals,"quoteid",(uint32_t)quoteid);
-        jaddnum(vals,"requestid",(uint32_t)requestid);
-        retstr = basilisk_standardservice("ACC",myinfo,0,myinfo->myaddr.persistent,vals,"",1);
-        free_json(vals);
-        return(retstr);
-    }
-}
-
-ZERO_ARGS(InstantDEX,init)
-{
-    basilisk_swaps_init(myinfo);
-    return(clonestr("{\"result\":\"success\"}"));
-}
-
-#include "../includes/iguana_apiundefs.h"
