+++ basilisk/basilisk_MSG.c
@@ -319,67 +319,6 @@ cJSON *dpow_addmessage(struct supernet_info *myinfo,char *jsonstr)
     return(retjson);
 }
 
-#include "../includes/iguana_apidefs.h"
-#include "../includes/iguana_apideclares.h"
-
-HASH_ARRAY_STRING(basilisk,getmessage,hash,vals,hexstr)
-{
-    uint32_t msgid,width,channel; char *retstr;
-    if ( bits256_cmp(GENESIS_PUBKEY,jbits256(vals,"srchash")) == 0 )
-        jaddbits256(vals,"srchash",hash);
-    if ( bits256_cmp(GENESIS_PUBKEY,jbits256(vals,"desthash")) == 0 )
-        jaddbits256(vals,"desthash",myinfo->myaddr.persistent);
-    if ( (msgid= juint(vals,"msgid")) == 0 )
-    {
-        msgid = (uint32_t)time(NULL);
-        jdelete(vals,"msgid");
-        jaddnum(vals,"msgid",msgid);
-    }
-    if ( myinfo->NOTARY.RELAYID >= 0 || myinfo->dexsock >= 0 || myinfo->subsock >= 0 )
-    {
-        channel = juint(vals,"channel");
-        width = juint(vals,"width");
-        retstr = basilisk_iterate_MSG(myinfo,channel,msgid,jbits256(vals,"srchash"),jbits256(vals,"desthash"),width);
-        //printf("getmessage.(%s)\n",retstr);
-        return(retstr);
-    }
-    //printf("getmessage not relay.%d dexsock.%d subsock.%d\n",myinfo->NOTARY.RELAYID,myinfo->dexsock,myinfo->subsock);
-    return(basilisk_standardservice("MSG",myinfo,0,jbits256(vals,"desthash"),vals,hexstr,1));
-}
-
-HASH_ARRAY_STRING(basilisk,sendmessage,hash,vals,hexstr)
-{
-    int32_t keylen,datalen,allocsize = 65536; uint8_t key[BASILISK_KEYSIZE],*space,*space2,*data,*ptr = 0; char *retstr=0;
-    space = calloc(1,allocsize);
-    space2 = calloc(1,allocsize);
-    data = get_dataptr(BASILISK_HDROFFSET,&ptr,&datalen,&space[BASILISK_KEYSIZE],allocsize-BASILISK_KEYSIZE,hexstr);
-    if ( myinfo->subsock >= 0 || myinfo->dexsock >= 0 || (myinfo->IAMNOTARY != 0 && myinfo->NOTARY.RELAYID >= 0) )
-    {
-        keylen = basilisk_messagekey(key,juint(vals,"channel"),juint(vals,"msgid"),jbits256(vals,"srchash"),jbits256(vals,"desthash"));
-        if ( data != 0 )
-        {
-            retstr = basilisk_respond_addmessage(myinfo,key,keylen,data,datalen,0,juint(vals,"duration"));
-        } else printf("no get_dataptr\n");
-        if ( retstr != 0 )
-            free(retstr);
-    } //else printf("not notary.%d relayid.%d\n",myinfo->IAMNOTARY,myinfo->NOTARY.RELAYID);
-    if ( vals != 0 && juint(vals,"fanout") == 0 )
-        jaddnum(vals,"fanout",MAX(8,(int32_t)sqrt(myinfo->NOTARY.NUMRELAYS)+2));
-    if ( BASILISK_KEYSIZE+datalen < allocsize )
-    {
-        memcpy(space2,key,BASILISK_KEYSIZE);
-        if ( data != 0 && datalen != 0 )
-            memcpy(&space2[BASILISK_KEYSIZE],data,datalen);
-        dex_reqsend(myinfo,"DEX",space2,datalen+BASILISK_KEYSIZE,1,"");
-    } else printf("sendmessage space too small error for %d\n",datalen);
-    free(space);
-    free(space2);
-    if ( ptr != 0 )
-        free(ptr);
-    return(basilisk_standardservice("OUT",myinfo,0,jbits256(vals,"desthash"),vals,hexstr,0));
-}
-#include "../includes/iguana_apiundefs.h"
-
 int32_t basilisk_channelsend(struct supernet_info *myinfo,bits256 srchash,bits256 desthash,uint32_t channel,uint32_t msgid,uint8_t *data,int32_t datalen,uint32_t duration)
 {
     char *retstr,*hexstr,strbuf[4096],*ptr = 0; int32_t retval = -1; cJSON *valsobj;
