+++ basilisk/basilisk_bitcoin.c
@@ -1049,175 +1049,6 @@ cJSON *BTC_makeclaimfunc(struct supernet_info *myinfo,struct exchange_info *exch
 }
 #endif
 
-
-#include "../includes/iguana_apidefs.h"
-#include "../includes/iguana_apideclares.h"
-
-HASH_ARRAY_STRING(basilisk,value,hash,vals,hexstr)
-{
-    char *retstr=0,*symbol,*coinaddr,*infostr; cJSON *retjson,*sobj,*info,*addrs,*txoutjson,*txjson,*array; uint32_t basilisktag,blocktime; bits256 txid,blockhash; struct basilisk_item *ptr,Lptr; uint64_t value; int32_t timeoutmillis,vout,height,n,m;
-    if ( vals == 0 )
-        return(clonestr("{\"error\":\"null valsobj\"}"));
-    //if ( myinfo->IAMNOTARY != 0 || myinfo->NOTARY.RELAYID >= 0 )
-    //    return(clonestr("{\"error\":\"special relays only do OUT and MSG\"}"));
-    //if ( coin == 0 )
-    {
-        if ( (symbol= jstr(vals,"symbol")) != 0 || (symbol= jstr(vals,"coin")) != 0 )
-            coin = iguana_coinfind(symbol);
-    }
-    if ( jobj(vals,"fanout") == 0 )
-        jaddnum(vals,"fanout",MAX(5,(int32_t)sqrt(myinfo->NOTARY.NUMRELAYS)+1));
-    txid = jbits256(vals,"txid");
-    vout = jint(vals,"vout");
-    if ( coin != 0 )
-    {
-        if ( coin->FULLNODE < 0 )
-        {
-            if ( (txoutjson= dpow_gettxout(myinfo,coin,txid,vout)) != 0 )
-            {
-                if ( (coinaddr= jstr(txoutjson,"address")) != 0 && (value= SATOSHIDEN*jdouble(txoutjson,"value")) != 0 )
-                {
-                    retjson = cJSON_CreateObject();
-                    jaddstr(retjson,"result","success");
-                    jaddstr(retjson,"address",coinaddr);
-                    jadd64bits(retjson,"satoshis",value);
-                    jaddnum(retjson,"value",dstr(value));
-                    jaddnum(retjson,"amount",dstr(value));
-                    height = dpow_getchaintip(myinfo,&blockhash,&blocktime,0,0,coin);
-                    jaddnum(retjson,"height",height);
-                    jaddnum(retjson,"numconfirms",jint(txoutjson,"confirmations"));
-                    jaddbits256(retjson,"txid",txid);
-                    jaddnum(retjson,"vout",vout);
-                    jaddstr(retjson,"coin",coin->symbol);
-                }
-                else
-                {
-                    free_json(txoutjson);
-                    return(clonestr("{\"error\":\"return from gettxout missing fields\"}"));
-                }
-                free_json(txoutjson);
-                return(jprint(retjson,1));
-            } else return(clonestr("{\"error\":\"null return from gettxout\"}"));
-        }
-        if ( (basilisktag= juint(vals,"basilisktag")) == 0 )
-            basilisktag = rand();
-        if ( (timeoutmillis= juint(vals,"timeout")) <= 0 )
-            timeoutmillis = BASILISK_TIMEOUT;
-        if ( coin->FULLNODE > 0 && (ptr= basilisk_bitcoinvalue(&Lptr,myinfo,coin,remoteaddr,basilisktag,timeoutmillis,vals)) != 0 )
-        {
-            retstr = ptr->retstr, ptr->retstr = 0;
-            ptr->finished = OS_milliseconds() + 10000;
-            return(retstr);
-        }
-    }
-    if ( myinfo->reqsock >= 0 )
-    {
-        if ( (retstr= _dex_getrawtransaction(myinfo,symbol,txid)) != 0 )
-        {
-            if ( (txoutjson= cJSON_Parse(retstr)) != 0 )
-            {
-                //printf("TX.(%s)\n",jprint(txoutjson,0));
-                retjson = cJSON_CreateObject();
-                jaddstr(retjson,"result","success");
-                jaddnum(retjson,"numconfirms",jint(txoutjson,"confirmations"));
-                if ( (array= jarray(&n,txoutjson,"vout")) != 0 && vout < n && (txjson= jitem(array,vout)) != 0 )
-                {
-                    //printf("txjson.(%s)\n",jprint(txjson,0));
-                    if ( (value= jdouble(txjson,"value") * SATOSHIDEN) != 0 )
-                    {
-                        if ( (sobj= jobj(txjson,"scriptPubKey")) != 0 && (addrs= jarray(&m,sobj,"addresses")) != 0 && (coinaddr= jstri(addrs,0)) != 0 )
-                            jaddstr(retjson,"address",coinaddr);
-                        jadd64bits(retjson,"satoshis",value);
-                        jaddnum(retjson,"value",dstr(value));
-                        if ( (infostr= _dex_getinfo(myinfo,symbol)) != 0 )
-                        {
-                            if ( (info= cJSON_Parse(infostr)) != 0 )
-                            {
-                                if ( (height= jint(info,"blocks")) > 0 )
-                                {
-                                    height -= jint(txoutjson,"confirmations");
-                                    jaddnum(retjson,"height",height);
-                                }
-                                free_json(info);
-                            }
-                            free(infostr);
-                        }
-                        jaddbits256(retjson,"txid",txid);
-                        jaddnum(retjson,"vout",vout);
-                        jaddstr(retjson,"coin",symbol);
-                        free(retstr);
-                        free_json(txoutjson);
-                        return(jprint(retjson,1));
-                    }
-                }
-                free_json(txoutjson);
-                return(jprint(retjson,1));
-            }
-            return(retstr);
-        }
-    }
-    return(basilisk_standardservice("VAL",myinfo,0,hash,vals,hexstr,1));
-}
-
-HASH_ARRAY_STRING(basilisk,rawtx,hash,vals,hexstr)
-{
-    char *retstr=0,*symbol; uint32_t basilisktag; int32_t timeoutmillis,i,retval = -1; uint64_t amount,txfee; cJSON *retarray;
-    if ( vals == 0 )
-        return(clonestr("{\"error\":\"null valsobj\"}"));
-    //if ( coin == 0 )
-    {
-        if ( (symbol= jstr(vals,"symbol")) != 0 || (symbol= jstr(vals,"coin")) != 0 )
-            coin = iguana_coinfind(symbol);
-    }
-    if ( jobj(vals,"numrequired") == 0 )
-        jaddnum(vals,"numrequired",MIN(3,(int32_t)sqrt(myinfo->NOTARY.NUMRELAYS)+1));
-    if ( jobj(vals,"fanout") == 0 )
-        jaddnum(vals,"fanout",MAX(3,(int32_t)sqrt(myinfo->NOTARY.NUMRELAYS)+1));
-    if ( coin != 0 )
-    {
-        //if ( juint(vals,"burn") == 0 )
-        //    jaddnum(vals,"burn",0.0001);
-        if ( (basilisktag= juint(vals,"basilisktag")) == 0 )
-            basilisktag = rand();
-        if ( (timeoutmillis= juint(vals,"timeout")) <= 0 )
-            timeoutmillis = BASILISK_TIMEOUT;
-        if ( (retstr= basilisk_bitcoinrawtx(myinfo,coin,remoteaddr,basilisktag,timeoutmillis,vals,0)) != 0 )
-        {
-            printf("rawtx.(%s)\n",retstr);
-            if ( (amount= j64bits(vals,"satoshis")) == 0 )
-                amount = jdouble(vals,"value") * SATOSHIDEN;
-            if ( (txfee= j64bits(vals,"txfee")) == 0 )
-                txfee = coin->chain->txfee;
-            if ( txfee == 0 )
-                txfee = 10000;
-            retval = -1;
-            if ( (retarray= cJSON_Parse(retstr)) != 0 )
-            {
-                if ( is_cJSON_Array(retarray) != 0 )
-                {
-                    for (i=0; i<cJSON_GetArraySize(retarray); i++)
-                    {
-                        if ( basilisk_vins_validate(myinfo,coin,jitem(retarray,i),amount,txfee) == 0 )
-                        {
-                            retval = 0;
-                            break;
-                        }
-                    }
-                } else retval = basilisk_vins_validate(myinfo,coin,retarray,amount,txfee);
-                if ( retval < 0 )
-                {
-                    printf("ERROR.(%s)\n",retstr);
-                    free(retstr);
-                    retstr = clonestr("{\"error\":\"invalid vin in rawtx\"}");
-                }
-            }
-        }
-    } else retstr = clonestr("{\"error\":\"no coin specified or found\"}");
-    return(retstr);
-}
-
-#include "../includes/iguana_apiundefs.h"
-
 int32_t basilisk_unspentfind(struct supernet_info *myinfo,struct iguana_info *coin,bits256 *txidp,int32_t *voutp,uint8_t *spendscript,struct iguana_outpoint outpt,int64_t value)
 {
     struct iguana_waccount *wacct,*tmp; struct iguana_waddress *waddr,*tmp2;
