+++ basilisk/tradebots_liquidity.c
@@ -1238,4 +1238,94 @@ ZERO_ARGS(tradebot,openliquidity)
     return(jprint(array,1));
 }
 
+ZERO_ARGS(tradebot,allbalances)
+{
+    int32_t i,n; double value,pending; char *base; cJSON *item,*balances = cJSON_CreateObject();
+    if ( myinfo->liquidity_currencies == 0 )
+        myinfo->liquidity_currencies = cJSON_Parse("[\"KMD\", \"BTC\"]");
+    if ( myinfo->liquidity_currencies != 0 && (n= cJSON_GetArraySize(myinfo->liquidity_currencies)) > 0 )
+    {
+        for (i=0; i<n; i++)
+        {
+            base = jstri(myinfo->liquidity_currencies,i);
+            value = tradebot_balance(myinfo,base);
+            pending = tradebot_pending(myinfo,base);
+            item = cJSON_CreateObject();
+            jaddnum(item,"value",value);
+            jaddnum(item,"pending",pending);
+            jadd(balances,base,item);
+        }
+    }
+    return(jprint(balances,1));
+}
+
+ZERO_ARGS(tradebot,anchor)
+{
+    FILE *fp; char *anchorstr,fname[512]; cJSON *anchor; int32_t retval = -1;
+    if ( (anchorstr= tradebot_allbalances(myinfo,0,0,0)) != 0 )
+    {
+        if ( (anchor= cJSON_Parse(anchorstr)) != 0 )
+        {
+            if ( jobj(anchor,"error") == 0 )
+            {
+                sprintf(fname,"%s/anchor",GLOBAL_DBDIR), OS_compatible_path(fname);
+                if ( (fp= fopen(fname,"wb")) != 0 )
+                {
+                    if ( fwrite(anchorstr,1,strlen(anchorstr)+1,fp) == strlen(anchorstr)+1 )
+                        retval = 0;
+                    fclose(fp);
+                }
+            }
+        }
+        free(anchorstr);
+    }
+    if ( retval == 0 )
+        return(clonestr("{\"result\":\"success\"}"));
+    else return(clonestr("{\"error\":\"couldnt make anchor file\"}"));
+}
+
+ZERO_ARGS(tradebot,portfolio)
+{
+    char *currentstr,*anchorstr,fname[512]; long fsize; cJSON *current,*anchor=0,*portfolio=0;
+    if ( (currentstr= tradebot_allbalances(myinfo,0,0,0)) != 0 )
+    {
+        if ( (current= cJSON_Parse(currentstr)) != 0 )
+        {
+            sprintf(fname,"%s/anchor",GLOBAL_DBDIR), OS_compatible_path(fname);
+            if ( (anchorstr= OS_filestr(&fsize,fname)) != 0 )
+            {
+                anchor = cJSON_Parse(anchorstr);
+                free(anchorstr);
+            }
+            if ( anchor == 0 )
+                anchor = cJSON_Parse("{}");
+            portfolio = tradebot_balancesdiff(myinfo,current,anchor);
+            free_json(current);
+        }
+        free(currentstr);
+    }
+    if ( portfolio == 0 )
+        return(clonestr("{\"result\":\"success\"}"));
+    else return(jprint(portfolio,1));
+}
+
+ARRAY_OBJ_INT(tradebot,goals,currencies,vals,targettime)
+{
+    static bits256 zero; char *targetcoin; int32_t i,n;
+    if ( currencies != 0 && vals != 0 )
+    {
+        // init things so automatically updates refli.bid and refli.ask
+        // volume range with margin
+        // currency percentage value in BTC? target distribution, max percentage, min percentage`
+        // min price to sell, max price to buy, max volume
+        n = cJSON_GetArraySize(currencies);
+        for (i=0; i<n; i++)
+        {
+            targetcoin = jstri(currencies,i);
+            tradebot_liquidity_command(myinfo,targetcoin,zero,vals);
+        }
+        return(clonestr("{\"result\":\"success\"}"));
+    } else return(clonestr("{\"error\":\"no currencies or vals\"}"));
+}
+
 #include "../includes/iguana_apiundefs.h"
