+++ gecko/gecko.c
@@ -273,9 +273,6 @@ char *basilisk_respond_geckoget(struct supernet_info *myinfo,char *CMD,void *add
     } else return(clonestr("{\"error\":\"invalid geckoget type, mustbe (HDR or BLK or GTX)\"}"));
 }
 
-#include "../includes/iguana_apidefs.h"
-#include "../includes/iguana_apideclares.h"
-
 char *gecko_sendrawtransaction(struct supernet_info *myinfo,char *symbol,uint8_t *data,int32_t datalen,bits256 txid,cJSON *vals,char *signedtx)
 {
     char *retstr = 0; struct iguana_info *virt,*btcd = iguana_coinfind("BTCD");
@@ -290,48 +287,5 @@ char *gecko_sendrawtransaction(struct supernet_info *myinfo,char *symbol,uint8_t
     return(retstr);
 }
 
-HASH_ARRAY_STRING(basilisk,geckotx,hash,vals,hexstr)
-{
-    struct iguana_info *btcd; char *retstr=0,*symbol; uint8_t *data,*allocptr,space[4096]; int32_t datalen; bits256 txid;
-    if ( (btcd= iguana_coinfind("BTCD")) != 0 && (symbol= jstr(vals,"symbol")) != 0 )
-    {
-        if ( (data= get_dataptr(BASILISK_HDROFFSET,&allocptr,&datalen,space,sizeof(space),hexstr)) != 0 )
-        {
-            txid = bits256_doublesha256(0,data,datalen);
-            retstr = gecko_sendrawtransaction(myinfo,symbol,data,datalen,txid,vals,hexstr);
-        } else retstr = clonestr("{\"error\":\"no tx submitted\"}");
-        if ( allocptr != 0 )
-            free(allocptr);
-        if ( retstr == 0 )
-            retstr = clonestr("{\"error\":\"couldnt create geckotx\"}");
-        return(retstr);
-    } return(clonestr("{\"error\":\"need symbol and chain and BTCD to create new gecko tx\"}"));
-}
-
-HASH_ARRAY_STRING(basilisk,geckoblock,hash,vals,hexstr)
-{
-    return(clonestr("{\"error\":\"geckoblock is an internal reporting function\"}"));
-}
-
-HASH_ARRAY_STRING(basilisk,geckoheaders,hash,vals,hexstr)
-{
-    return(clonestr("{\"error\":\"geckoheaders is an internal reporting function\"}"));
-}
-
-HASH_ARRAY_STRING(basilisk,geckoget,hash,vals,hexstr)
-{
-    struct iguana_info *btcd,*virt; char *symbol;
-    if ( (btcd= iguana_coinfind("BTCD")) != 0 && (symbol= jstr(vals,"symbol")) != 0 )
-    {
-        if ( (virt= iguana_coinfind(symbol)) != 0 )
-        {
-            basilisk_wait(myinfo,virt);
-            return(basilisk_respond_geckoget(myinfo,"GET",&coin->internaladdr,remoteaddr,0,vals,0,0,hash,0));
-        } else return(clonestr("{\"error\":\"geckoget needs virtualchain\"}"));
-    }
-    return(clonestr("{\"error\":\"geckoget needs BTCD\"}"));
-}
-
-#include "../includes/iguana_apiundefs.h"
