+++ iguana/iguana_bitmap.c
@@ -1162,42 +1162,3 @@ void iguana_bitmap(char *space,int32_t max,char *name)
     }
 }
 
-#include "../includes/iguana_apidefs.h"
-#include "../includes/iguana_apideclares.h"
-
-STRING_AND_TWOINTS(mouse,change,name,x,y)
-{
-    printf("mouse (%s) x.%d y.%d\n",name,x,y);
-    return(clonestr("{\"result\":\"changed\"}"));
-}
-
-STRING_ARG(mouse,leave,name)
-{
-    printf("mouse (%s) leave\n",name);
-    return(clonestr("{\"result\":\"left\"}"));
-}
-
-STRING_AND_TWOINTS(mouse,click,name,x,y)
-{
-    printf("mouse (%s) x.%d y.%d click\n",name,x,y);
-    return(clonestr("{\"result\":\"click\"}"));
-}
-
-STRING_AND_INT(keyboard,key,name,c)
-{
-    printf(" KEY.(%s) c.%d (%c)\n",name,c,c);
-    return(clonestr("{\"result\":\"key\"}"));
-}
-
-STRING_AND_TWOINTS(mouse,image,name,x,y)
-{
-    printf("mouse CREATE (%s) x.%d y.%d\n",name,x,y);
-    return(clonestr("{\"result\":\"opened\"}"));
-}
-
-STRING_ARG(mouse,close,name)
-{
-    printf("mouse CLOSE (%s)\n",name);
-    return(clonestr("{\"result\":\"closed\"}"));
-}
-#include "../includes/iguana_apiundefs.h"
